[common-forms] reorganized files again
