+++ src/components/forms/forms.js
@@ -1,27 +1,26 @@
 import { useState, useEffect } from 'react';
-import { h, div, label, input, span, button } from 'react-hyperscript-helpers';
+import { h, div, label, span, button } from 'react-hyperscript-helpers';
 import { cloneDeep, isNil, isEmpty } from 'lodash/fp';
 import { SearchSelectOrText } from '../SearchSelectOrText';
 import Creatable from 'react-select/creatable';
 import { isEmailAddress } from '../../libs/utils';
+import {
+  formInputGeneric, formInputMultiText,
+  formInputSelectOrCreate, formInputSelect,
+  formInputCheckbox,
+  formInputSlider
+} from './formComponents';
 
 import './forms.css';
 
 export const FormFieldTypes = {
-  SELECT: { id: 'select', defaultValue: '' },
-  SELECT_CREATABLE: { id: 'selectCreatable', defaultValue: '' },
-  MULTITEXT: { id: 'multitext', defaultValue: [] },
-  CHECKBOX: { id: 'checkbox', defaultValue: false },
-  SLIDER: { id: 'slider', defaultValue: false },
-  TEXT: { id: 'text', defaultValue: '' },
-  NUMBER: { id: 'number', defaultValue: '' }
-};
-
-export const styles = {
-  inputStyle: {
-    padding: '25px 15px',
-    width: '100%'
-  }
+  SELECT: { defaultValue: '', component: formInputSelect },
+  SELECT_CREATABLE: { defaultValue: '', component: formInputSelectOrCreate },
+  MULTITEXT: { defaultValue: [], component: formInputMultiText },
+  CHECKBOX: { defaultValue: false, component: formInputCheckbox },
+  SLIDER: { defaultValue: false, component: formInputSlider },
+  TEXT: { defaultValue: '', component: formInputGeneric },
+  NUMBER: { defaultValue: '', component: formInputGeneric },
 };
 
 export const FormValidators = {
@@ -45,298 +44,6 @@ export const FormValidators = {
     msg: 'Please enter a valid email address (e.g., person@example.com)'
   }
 };
-
-//---------------------------------------------
-// Form Behavior
-//---------------------------------------------
-const validateFormInput = (config, value) => {
-  const { setError, validators } = config;
-  if (validators) {
-    const validationResults = validators
-      .filter(validator => !validator.isValid(value))
-      .map(x => x.msg);
-
-    const isValid = validationResults.length === 0;
-    setError(isValid ? undefined : validationResults);
-    return isValid;
-  }
-  setError();
-  return true;
-};
-
-const normalizeValue = (value) => {
-  if (typeof value === 'string') {
-    return value.trim();
-  } else if (Array.isArray(value)) {
-    return value
-      .map(x => normalizeValue(x))
-      .filter(x => x); // filter out null strings
-  }
-  return value;
-};
-
-const onFormInputChange = (config, value) => {
-  const { id, onChange, setFormValue } = config;
-  const normalizedValue = normalizeValue(value);
-  const isValidInput = validateFormInput(config, normalizedValue);
-
-  onChange({key: id, value: normalizedValue, isValid: isValidInput });
-  setFormValue(value);
-};
-
-const errorMessage = (error) => {
-  return error && div({ className: `error-message fadein`}, [
-    span({ className: 'glyphicon glyphicon-play' }),
-    ...error.map((err) => div([err])),
-  ]);
-};
-
-//---------------------------------------------
-// Form Controls
-//---------------------------------------------
-const formInput = (config) => {
-  switch (config.type) {
-    case FormFieldTypes.SELECT: return formInputSelect(config);
-    case FormFieldTypes.SELECT_CREATABLE: return formInputSelectOrCreate(config);
-    case FormFieldTypes.MULTITEXT: return formInputMultiText(config);
-    case FormFieldTypes.CHECKBOX: return formInputCheckbox(config);
-    case FormFieldTypes.SLIDER: return formInputSlider(config);
-    case FormFieldTypes.TEXT:
-    case FormFieldTypes.NUMBER:
-    default:
-      return formInputGeneric(config);
-  }
-};
-
-const formInputGeneric = (config) => {
-  const {
-    id, title, type, disabled,
-    placeholder,
-    inputStyle, ariaDescribedby,
-    formValue, error, setError
-  } = config;
-
-  return div([
-    input({
-      id: id,
-      type: type || 'text',
-      className: `form-control ${error ? 'errored' : ''}`,
-      placeholder: placeholder || title,
-      value: formValue,
-      style: { ...styles.inputStyle, ...inputStyle },
-      disabled: disabled,
-      onChange: (event) => onFormInputChange(config, event.target.value),
-      onFocus: () => setError(),
-      onBlur: (event) => validateFormInput(config, event.target.value),
-      'aria-describedby': ariaDescribedby,
-    }),
-    errorMessage(error)
-  ]);
-};
-
-const formInputMultiText = (config) => {
-  const {
-    id, title, disabled,
-    placeholder, ariaDescribedby,
-    inputStyle, onChange,
-    formValue, setFormValue, error, setError
-  } = config;
-
-  const pushValue = (element) => {
-    const value = element.value.trim();
-
-    if (!value || !validateFormInput(config, value)) {
-      return;
-    }
-    if (formValue.indexOf(value) !== -1) {
-      element.value = '';
-      return;
-    }
-
-    const formValueClone = cloneDeep(formValue);
-    formValueClone.push(value);
-    setFormValue(formValueClone);
-    onChange({key: id, value: formValueClone, isValid: true});
-    element.value = '';
-  };
-
-  const removePill = (index) => {
-    const formValueClone = cloneDeep(formValue);
-    formValueClone.splice(index, 1);
-    setFormValue(formValueClone);
-    onChange({key: id, value: formValueClone, isValid: true});
-  };
-
-  return div({}, [
-    div({
-      className: 'formControl-group flex-row',
-    }, [
-      input({
-        id,
-        type: 'text',
-        className: `form-control ${error ? 'errored' : ''}`,
-        placeholder: placeholder || title,
-        style: { ...styles.inputStyle, ...inputStyle },
-        disabled,
-        'aria-describedby': ariaDescribedby,
-        onKeyUp: (event) => event.code === 'Enter' ? pushValue(event.target) : setError(),
-        onFocus: () => setError()
-      }),
-      h(button, {
-        className: 'form-btn btn-xs',
-        type: 'button',
-        disabled,
-        style: {
-          marginTop: 0,
-          minWidth: 'fit-content'
-        },
-        onClick: () => pushValue(document.getElementById(id))
-      }, [
-        span({
-          className: 'glyphicon glyphicon-plus',
-          'aria-label': 'Add',
-          style: { margin: '0 8px' },
-          isRendered: !disabled
-        })
-      ])
-    ]),
-    errorMessage(error),
-    div({ className: 'flex-row', style: { justifyContent: null } },
-      formValue.map((val, i) => {
-        return h(button, {
-          key: val,
-          className: 'pill btn-xs',
-          type: 'button',
-          disabled,
-          onClick: () => removePill(i)
-        }, [
-          val,
-          span({
-            className: 'glyphicon glyphicon-remove',
-            style: { marginLeft: '8px' },
-            isRendered: !disabled
-          })
-        ]);
-      })
-    )
-  ]);
-};
-
-// Using react-select/creatable - Passing config directly through!
-const formInputSelectOrCreate = (config) => {
-  const {
-    id, title, disabled, required, error, setError,
-    selectOptions, searchPlaceholder, ariaDescribedby,
-    formValue,
-    creatableConfig = {}
-  } = config;
-
-  return h(Creatable, {
-    key: id,
-    isClearable: true, //ensures that selections can be cleared from dropdown, adds an 'x' within input box
-    required,
-    isDisabled: disabled,
-    placeholder: searchPlaceholder || `Search for ${title}...`,
-    className: `form-select ${error ? 'errored' : ''}`,
-    onChange: (option) => onFormInputChange(config, option),
-    onMenuOpen: () => setError(),
-    onMenuClose: () => {
-      if (required && !formValue) {
-        setError(FormValidators.REQUIRED.msg);
-      }
-    },
-    options: selectOptions,
-    getOptionLabel: (option) => option.displayValue,
-    getNewOptionData: (inputValue) => {
-      return { displayValue: inputValue };
-    },
-    getOptionValue: (option) => { //value formatter for options, attr used to ensure empty strings are treated as undefined
-      if(isNil(option) || isEmpty(option.displayName)) {
-        return null;
-      }
-      return option;
-    },
-    value: formValue,
-    ...creatableConfig,
-    'aria-describedby': ariaDescribedby
-  });
-};
-
-const formInputSelect = (config) => {
-  const {
-    id, title, disabled, error, setError,
-    selectOptions, searchPlaceholder, ariaDescribedby
-  } = config;
-
-  return div({}, [
-    h(SearchSelectOrText, {
-      id,
-      'aria-describedby': ariaDescribedby,
-      onPresetSelection: async (selection) => onFormInputChange(config, selection),
-      onManualSelection: (selection) => onFormInputChange(config, selection),
-      onOpen: () => setError(),
-      options: selectOptions.map((x) => {
-        return typeof x == 'string' || typeof x === 'number'
-          ? { key: x, displayText: x }
-          : x;
-      }),
-      searchPlaceholder: searchPlaceholder || `Search for ${title}...`,
-      className: 'form-control',
-      disabled, errored: error
-    }),
-    errorMessage(error)
-  ]);
-};
-
-const formInputCheckbox = (config) => {
-  const {
-    id, disabled, error, toggleText,
-    formValue, ariaDescribedby
-  } = config;
-
-  return div({ className: 'checkbox' }, [
-    input({
-      type: 'checkbox',
-      id: `cb_${id}_${toggleText}`,
-      checked: formValue,
-      className: 'checkbox-inline',
-      'aria-describedby': ariaDescribedby,
-      onChange: (event) => onFormInputChange(config, event.target.checked),
-      disabled
-    }),
-    label({
-      className: `regular-checkbox ${error ? 'errored' : ''}`,
-      htmlFor: `cb_${id}_${toggleText}`,
-    }, [toggleText])
-  ]);
-};
-
-const formInputSlider = (config) => {
-  const {
-    id, disabled, toggleText, formValue
-  } = config;
-
-  return div({ className: 'flex-row', style: { justifyContent: 'unset' } }, [
-    label({ className: 'switch', htmlFor: `cb_${id}_${toggleText}` }, [
-      input({
-        type: 'checkbox',
-        id: `cb_${id}_${toggleText}`,
-        checked: formValue,
-        className: 'checkbox-inline',
-        onChange: (event) => onFormInputChange(config, event.target.checked),
-        disabled
-      }),
-      div({className: 'slider round'}),
-    ]),
-    div({
-      style: {
-        marginLeft: 15,
-        fontStyle: 'italic'
-      }
-    }, [toggleText])
-  ]);
-};
-
 
 //---------------------------------------------
 // Main Components
@@ -366,7 +73,7 @@ const formInputSlider = (config) => {
 */
 export const FormField = (config) => {
   const {
-    id, type, ariaLevel,
+    id, type = FormFieldTypes.TEXT, ariaLevel,
     title, hideTitle, description,
     defaultValue, style, validators
   } = config;
@@ -396,7 +103,7 @@ export const FormField = (config) => {
       required && '*'
     ]),
     description && div({ style: { marginBottom: 15 } }, description),
-    formInput({
+    h(type.component, {
       ...config,
       error, setError,
       formValue, setFormValue,
